[VSC-9] Reorganize module structure to do tests
