+++ src/server/test/helpers.ts
@@ -1,0 +1,2 @@
+export const wait = (s: number) =>
+    new Promise((resolve) => setTimeout(resolve, s * 1000));
